mcps/spotify/pyproject.toml
[project]
name = "spotify-mcp-server"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.0.2",
    "python-dotenv>=1.0",
    "spotipy>=2.26.0",
]
